requirements.txt
python-docx>=1.1.0
PyPDF2>=3.0.0
flask>=3.0.0
